[FIX] make getVolcanoPlotData smart enough to get its own activity data
Previously, this method would generate an error if related activity data were
missing. This was okay because we were always generating a heatmap, which would
load that data in advance, so this was really an error. Now that we have
separated the heatmap out of the sampleBin, we can no longer rely on that
mechanism. This change makes the error condition into a call that will load any
missing data before we proceed.